Implementing duplicate cabin

+++ src/services/apiCabins.js
@@ -39,11 +39,13 @@ export async function createEditCabin(newCabin, id) {
   }
 
   // 2. Upload image
+  if (hasImagePath) return data;
+
   const { error: storageError } = await supabase.storage
     .from("cabin-images")
     .upload(imageName, newCabin.image);
 
-  // 3. Delete the cabin IF there wae an error uploading image
+  // 3. Delete the cabin IF there was an error uploading image
   if (storageError) {
     await supabase.from("cabins").delete().eq("id", data.id);
     console.error(storageError);
